delete still failing

+++ backend/flaskr/__init__.py
@@ -121,14 +121,14 @@ def create_app(test_config=None):
     This removal will persist in the database and when you refresh the page.
     """
     
-    @app.route('/questions/<int:question_id>', methods=['DELETE'])
+    @app.route("/questions/<int:question_id>", methods=["DELETE"])
     def delete_question(question_id):
         try:
             question = Question.query.get(question_id)
             
-            # if not question:
-            #     abort(404)
-            # Question.delete(question)
+            if not question:
+                abort(404)
+
             question.delete()
             
             return jsonify({
@@ -136,7 +136,7 @@ def create_app(test_config=None):
                 'deleted': question_id
             })
         except:
-            abort(404)
+            abort(422)
 
     """
     @TODO:
